refactor(workspace): drop packages/workflows + core/agents, consolidate @core/sse
Inline launchDBOS into apps/worker (single file boot), delete obsolete
packages/workflows and core/agents (Rubi stub, unused models). Collapse
@core/sse split (channels/define/publisher/subscriber/types) into a single
core/sse/src/sse.ts. Drop classify.ts single-tx (batch only) and rename
keyword-match.ts → utils.ts. Remove jobPublisher mock from oRPC test ctx.
