+++ apps/web/package.json
@@ -21,6 +21,7 @@
       "@core/logging": "workspace:*",
       "@core/posthog": "workspace:*",
       "@core/redis": "workspace:*",
+      "@core/sse": "workspace:*",
       "@core/stripe": "workspace:*",
       "@core/utils": "workspace:*",
       "@dbos-inc/dbos-sdk": "catalog:workers",
@@ -51,7 +52,6 @@
       "@packages/analytics": "workspace:*",
       "@packages/events": "workspace:*",
       "@packages/ui": "workspace:*",
-      "@packages/workflows": "workspace:*",
       "@tailwindcss/vite": "catalog:vite",
       "@tanstack/ai": "catalog:tanstack-ai",
       "@tanstack/ai-openrouter": "catalog:tanstack-ai",
