+++ apps/web/tsconfig.json
@@ -19,15 +19,13 @@
          "@core/orpc/*": ["../../core/orpc/src/*"],
          "@core/posthog/*": ["../../core/posthog/src/*"],
          "@core/redis/*": ["../../core/redis/src/*"],
+         "@core/sse": ["../../core/sse/src/sse.ts"],
          "@core/stripe": ["../../core/stripe/src/index.ts"],
          "@core/stripe/*": ["../../core/stripe/src/*"],
          "@core/transactional/*": ["../../core/transactional/src/*"],
          "@core/utils/*": ["../../core/utils/src/*"],
-         "@core/agents": ["../../core/agents/src/mastra/index.ts"],
-         "@core/agents/*": ["../../core/agents/src/*"],
          "@packages/analytics/*": ["../../packages/analytics/src/*"],
          "@packages/events/*": ["../../packages/events/src/*"],
-         "@packages/workflows/*": ["../../packages/workflows/src/*"],
          "@packages/ui/*": ["../../packages/ui/src/*"],
          "@modules/billing/*": ["../../modules/billing/src/*"],
          "@modules/classification/*": ["../../modules/classification/src/*"],
@@ -41,9 +39,6 @@
    "references": [
       {
          "path": "../../core/transactional"
-      },
-      {
-         "path": "../../packages/workflows"
       },
       {
          "path": "../../packages/ui"
