+++ apps/worker/package.json
@@ -18,8 +18,7 @@
       "@core/transactional": "workspace:*",
       "@dbos-inc/dbos-sdk": "catalog:workers",
       "@modules/billing": "workspace:*",
-      "@modules/classification": "workspace:*",
-      "@packages/workflows": "workspace:*"
+      "@modules/classification": "workspace:*"
    },
    "devDependencies": {
       "@tooling/typescript": "workspace:*",
