+++ apps/worker/src/index.ts
@@ -1,3 +1,4 @@
+import { DBOS } from "@dbos-inc/dbos-sdk";
 import { env } from "@core/environment/worker";
 import { initLogger, getLogger } from "@core/logging/root";
 import { initOtel, shutdownOtel } from "@core/logging/otel";
@@ -5,7 +6,6 @@ import { createDb } from "@core/database/client";
 import { createRedis } from "@core/redis/connection";
 import { createPostHog } from "@core/posthog/server";
 import { createResendClient } from "@core/transactional/utils";
-import { launchDBOS } from "@packages/workflows/setup";
 import { setupBillingWorkflows } from "@modules/billing/workflows/setup";
 import { setupClassificationWorkflows } from "@modules/classification/workflows/setup";
 
@@ -33,17 +33,34 @@ const classification = await setupClassificationWorkflows({
    workerConcurrency: 10,
 });
 
-launchDBOS({
-   db,
-   redis,
-   posthog,
-   resendClient,
+DBOS.setConfig({
+   name: "montte-worker",
    systemDatabaseUrl: env.DATABASE_URL,
-   logLevel: env.LOG_LEVEL,
-   onLaunch: classification.applySchedules,
-   onShutdown: async () => {
-      await posthog.shutdown();
-      redis.disconnect();
-      await shutdownOtel();
-   },
+   logLevel: env.LOG_LEVEL ?? "info",
+   runAdminServer: false,
 });
+
+DBOS.launch()
+   .then(async () => {
+      logger.info("DBOS runtime started");
+      await classification.applySchedules();
+   })
+   .catch((err: unknown) => {
+      logger.error({ err }, "DBOS launch failed");
+   });
+
+async function gracefulShutdown(signal: string) {
+   logger.info(`${signal} received — shutting down`);
+   await DBOS.shutdown();
+   await posthog.shutdown();
+   redis.disconnect();
+   await shutdownOtel();
+   logger.info("Shutdown complete");
+   process.exit(0);
+}
+
+process.on("SIGTERM", () => void gracefulShutdown("SIGTERM"));
+process.on("SIGINT", () => void gracefulShutdown("SIGINT"));
+
+// db is required for module setups via singletons
+void db;
