+++ apps/worker/tsconfig.json
@@ -4,8 +4,6 @@
       "composite": false,
       "noEmit": true,
       "paths": {
-         "@core/agents": ["../../core/agents/src/agent.ts"],
-         "@core/agents/*": ["../../core/agents/src/*"],
          "@core/database": ["../../core/database/src/index.ts"],
          "@core/database/*": ["../../core/database/src/*"],
          "@core/environment/worker": ["../../core/environment/src/worker.ts"],
@@ -19,15 +17,11 @@
          "@core/transactional/*": ["../../core/transactional/dist/*"],
          "@modules/billing/*": ["../../modules/billing/src/*"],
          "@modules/classification/*": ["../../modules/classification/src/*"],
-         "@packages/workflows/*": ["../../packages/workflows/src/*"],
          "@packages/events/*": ["../../packages/events/src/*"]
       }
    },
    "include": ["src"],
    "references": [
-      {
-         "path": "../../packages/workflows"
-      },
       {
          "path": "../../core/stripe"
       },
