--- core/agents/.oxlintrc.json
@@ -1,1 +1,0 @@
-{ "extends": ["../../tooling/oxc/core.json"] }
