--- core/agents/oxlint.json
@@ -1,4 +1,0 @@
-{
-   "$schema": "../../node_modules/oxlint/configuration_schema.json",
-   "extends": ["../../tooling/oxc/core.json"]
-}
