--- core/agents/package.json
@@ -1,44 +1,0 @@
-{
-   "name": "@core/agents",
-   "version": "0.26.0",
-   "private": true,
-   "license": "Apache-2.0",
-   "files": [
-      "dist"
-   ],
-   "type": "module",
-   "exports": {
-      ".": {
-         "types": "./dist/agent.d.ts",
-         "default": "./dist/agent.js"
-      },
-      "./models": {
-         "types": "./dist/models.d.ts",
-         "default": "./dist/models.js"
-      },
-      "./utils": {
-         "types": "./dist/utils.d.ts",
-         "default": "./dist/utils.js"
-      }
-   },
-   "scripts": {
-      "build": "tsc --build",
-      "check": "oxlint ./src",
-      "typecheck": "tsgo"
-   },
-   "dependencies": {
-      "@core/database": "workspace:*",
-      "@core/environment": "workspace:*",
-      "@core/logging": "workspace:*",
-      "@core/posthog": "workspace:*",
-      "@core/utils": "workspace:*",
-      "@tanstack/ai": "catalog:tanstack-ai",
-      "@tanstack/ai-openrouter": "catalog:tanstack-ai",
-      "neverthrow": "catalog:validation",
-      "zod": "catalog:validation"
-   },
-   "devDependencies": {
-      "@tooling/typescript": "workspace:*",
-      "typescript": "catalog:development"
-   }
-}
