--- core/agents/src/agent.ts
@@ -1,8 +1,0 @@
-export type ChatRubiOptions = {
-   modelId?: string;
-   language?: string;
-};
-
-export function chatRubi(_options: ChatRubiOptions): never {
-   throw new Error("Not implemented — Rubi chat agent coming soon");
-}
