--- core/agents/tsconfig.json
@@ -1,40 +1,0 @@
-{
-   "extends": "@tooling/typescript/tsconfig.package.json",
-   "compilerOptions": {
-      "paths": {
-         "@core/agents": ["./src/agent.ts"],
-         "@core/agents/*": ["./src/*"],
-         "@core/authentication/*": ["../authentication/src/*"],
-         "@core/database": ["../database/src/index.ts"],
-         "@core/database/*": ["../database/src/*"],
-         "@core/environment/*": ["../environment/src/*"],
-         "@core/files/*": ["../files/src/*"],
-         "@core/logging": ["../logging/src/logger.ts"],
-         "@core/logging/*": ["../logging/src/*"],
-         "@core/posthog/*": ["../posthog/src/*"],
-         "@core/redis/*": ["../redis/src/*"],
-         "@core/stripe": ["../stripe/src/index.ts"],
-         "@core/stripe/*": ["../stripe/src/*"],
-         "@core/transactional/*": ["../transactional/src/*"],
-         "@core/utils/*": ["../utils/src/*"]
-      }
-   },
-   "include": ["src"],
-   "references": [
-      {
-         "path": "../utils"
-      },
-      {
-         "path": "../logging"
-      },
-      {
-         "path": "../environment"
-      },
-      {
-         "path": "../database"
-      },
-      {
-         "path": "../posthog"
-      }
-   ]
-}
