--- core/sse/src/channels.ts
@@ -1,5 +1,0 @@
-import type { SseScope } from "./types";
-
-export function channelFor(scope: SseScope): string {
-   return `sse:${scope.kind}:${scope.id}`;
-}
