+++ core/sse/tsconfig.json
@@ -2,7 +2,7 @@
    "extends": "@tooling/typescript/tsconfig.package.json",
    "compilerOptions": {
       "paths": {
-         "@core/sse/*": ["./src/*"],
+         "@core/sse": ["./src/sse.ts"],
          "@core/redis/*": ["../redis/src/*"]
       }
    },
