+++ modules/classification/tsconfig.json
@@ -17,7 +17,7 @@
          "@core/orpc/*": ["../../core/orpc/src/*"],
          "@core/posthog/*": ["../../core/posthog/src/*"],
          "@core/redis/*": ["../../core/redis/src/*"],
-         "@core/sse/*": ["../../core/sse/src/*"],
+         "@core/sse": ["../../core/sse/src/sse.ts"],
          "@core/stripe/*": ["../../core/stripe/src/*"],
          "@packages/events/*": ["../../packages/events/src/*"]
       }
