+++ packages/ui/tsconfig.json
@@ -16,8 +16,6 @@
          "@core/stripe/*": ["../../core/stripe/src/*"],
          "@core/transactional/*": ["../../core/transactional/src/*"],
          "@core/utils/*": ["../../core/utils/src/*"],
-         "@core/agents": ["../../core/agents/src/mastra/index.ts"],
-         "@core/agents/*": ["../../core/agents/src/*"],
          "@packages/analytics/*": ["../../packages/analytics/src/*"],
          "@packages/events/*": ["../../packages/events/src/*"],
          "@packages/ui/*": ["./src/*"]
