--- packages/workflows/package.json
@@ -1,43 +1,0 @@
-{
-   "name": "@packages/workflows",
-   "version": "0.1.0",
-   "private": true,
-   "license": "Apache-2.0",
-   "files": [
-      "dist"
-   ],
-   "type": "module",
-   "exports": {
-      "./setup": {
-         "types": "./dist/setup.d.ts",
-         "default": "./dist/setup.js"
-      }
-   },
-   "scripts": {
-      "build": "tsc --build",
-      "check": "oxlint ./src",
-      "format": "oxfmt --write ./src",
-      "format:check": "oxfmt --check ./src",
-      "typecheck": "tsgo"
-   },
-   "dependencies": {
-      "@core/agents": "workspace:*",
-      "@core/database": "workspace:*",
-      "@core/dbos": "workspace:*",
-      "@core/logging": "workspace:*",
-      "@core/posthog": "workspace:*",
-      "@core/redis": "workspace:*",
-      "@core/stripe": "workspace:*",
-      "@dbos-inc/dbos-sdk": "catalog:workers",
-      "@f-o-t/money": "catalog:fot",
-      "@packages/events": "workspace:*",
-      "@tanstack/store": "catalog:tanstack",
-      "dayjs": "catalog:ui",
-      "neverthrow": "catalog:validation",
-      "zod": "catalog:validation"
-   },
-   "devDependencies": {
-      "@tooling/typescript": "workspace:*",
-      "typescript": "catalog:development"
-   }
-}
