--- packages/workflows/tsconfig.json
@@ -1,50 +1,0 @@
-{
-   "extends": "@tooling/typescript/tsconfig.package.json",
-   "compilerOptions": {
-      "paths": {
-         "@core/agents": ["../../core/agents/src/agent.ts"],
-         "@core/agents/*": ["../../core/agents/src/*"],
-         "@core/database": ["../../core/database/src/index.ts"],
-         "@core/database/*": ["../../core/database/src/*"],
-         "@core/dbos/*": ["../../core/dbos/src/*"],
-         "@core/logging": ["../../core/logging/src/logger.ts"],
-         "@core/logging/*": ["../../core/logging/src/*"],
-         "@core/posthog/*": ["../../core/posthog/src/*"],
-         "@core/redis/*": ["../../core/redis/src/*"],
-         "@core/stripe": ["../../core/stripe/src/index.ts"],
-         "@core/stripe/*": ["../../core/stripe/src/*"],
-         "@packages/events/*": ["../../packages/events/src/*"],
-         "@packages/workflows/*": ["./src/*"]
-      }
-   },
-   "include": ["src"],
-   "references": [
-      {
-         "path": "../../core/transactional"
-      },
-      {
-         "path": "../../core/dbos"
-      },
-      {
-         "path": "../events"
-      },
-      {
-         "path": "../../core/stripe"
-      },
-      {
-         "path": "../../core/redis"
-      },
-      {
-         "path": "../../core/posthog"
-      },
-      {
-         "path": "../../core/logging"
-      },
-      {
-         "path": "../../core/database"
-      },
-      {
-         "path": "../../core/agents"
-      }
-   ]
-}
